+++ ScratchBasic/app/src/main/java/com/srjengbro/scratchbasic/operators/GreaterThanOperator.java
@@ -22,10 +22,10 @@ public class GreaterThanOperator extends Operator {
      * @return 1 if lhs is bigger otherwise 0
      */
     @Override
-    public Integer evaluate(Integer lhs, Integer rhs) {
+    public Double evaluate(Double lhs, Double rhs) {
         if (lhs > rhs) {
-            return 1;
+            return 1.0;
         }
-        return 0;
+        return 0.0;
     }
 }
